refactor: split 10 large source files into ~200-line modules
Refactored the following files to improve maintainability and code organization:

**Headers (3 files):**
- command.h: Split into 6 focused headers (command_base.h, command_chain.h,
  command_reorder.h, command_param.h, command_preset.h, command_clear.h) with
  umbrella include for backward compatibility

**Core Audio (7 files):**
- audio_engine.cpp: Split into audio_engine_chain.cpp, audio_engine_process.cpp,
  audio_engine_api.cpp (~100-175 lines each)
- audio_backend_portaudio.cpp: Split into audio_backend_portaudio_lifecycle.cpp,
  audio_backend_portaudio_devices.cpp, plus audio_backend_portaudio_helpers.h
  and audio_backend_portaudio_internal.h
- recorder.cpp: Split into recorder_session.cpp, recorder_io.cpp, recorder_wav.cpp,
  plus recorder_impl.h

**MIDI (1 file):**
- midi_manager.cpp: Split into midi_manager_mapping.cpp, midi_manager_persist.cpp

**GUI (6 files):**
- gui_manager.cpp: Split into gui_manager_frame.cpp, gui_manager_menu.cpp,
  gui_manager_update.cpp (nanosvg definitions remain in main file)
- file_dialog_native.cpp: Split into file_dialog_native_open.cpp,
  file_dialog_native_folder.cpp
- pedal_board.cpp: Split into pedal_board_menu.cpp, pedal_board_chain.cpp
- pedal_widget.cpp: Split into pedal_widget_body.cpp, pedal_widget_knobs.cpp

**Presets (1 file):**
- preset_manager.cpp: Split into preset_manager_dirs.cpp, preset_manager_io.cpp,
  plus preset_manager_impl.h

All files now target ~200 lines, improving readability and reducing cognitive load.
No API changes or breaking changes. All 175 tests pass. Both main executable and
test suite build successfully.

CMakeLists.txt updated with all new source files for all targets and platforms.
