+++ src/audio/audio_backend_portaudio_devices.cpp
@@ -1,0 +1,131 @@
+// =============================================================================
+// PortAudio backend — device management
+// =============================================================================
+
+#include "audio/audio_engine.h"
+#include "audio/audio_backend.h"
+#include "audio/audio_backend_portaudio_helpers.h"
+#include "audio/audio_backend_portaudio_internal.h"
+#include <portaudio.h>
+#include <iostream>
+#include <vector>
+
+namespace Amplitron {
+
+std::string AudioEngine::get_input_device_name() const {
+    if (input_device_ >= 0) {
+        const PaDeviceInfo* info = Pa_GetDeviceInfo(input_device_);
+        if (info) return info->name;
+    }
+    return "None";
+}
+
+std::string AudioEngine::get_output_device_name() const {
+    if (output_device_ >= 0) {
+        const PaDeviceInfo* info = Pa_GetDeviceInfo(output_device_);
+        if (info) return info->name;
+    }
+    return "None";
+}
+
+std::vector<AudioDeviceInfo> AudioEngine::get_input_devices() const {
+    std::vector<AudioDeviceInfo> devices;
+    int count = Pa_GetDeviceCount();
+    for (int i = 0; i < count; ++i) {
+        const PaDeviceInfo* info = Pa_GetDeviceInfo(i);
+        if (info && info->maxInputChannels > 0) {
+            devices.push_back({
+                i, info->name,
+                info->maxInputChannels, info->maxOutputChannels,
+                info->defaultSampleRate,
+                is_usb_device_name(info->name)
+            });
+        }
+    }
+    return devices;
+}
+
+std::vector<AudioDeviceInfo> AudioEngine::get_output_devices() const {
+    std::vector<AudioDeviceInfo> devices;
+    int count = Pa_GetDeviceCount();
+    for (int i = 0; i < count; ++i) {
+        const PaDeviceInfo* info = Pa_GetDeviceInfo(i);
+        if (info && info->maxOutputChannels > 0) {
+            devices.push_back({
+                i, info->name,
+                info->maxInputChannels, info->maxOutputChannels,
+                info->defaultSampleRate,
+                is_usb_device_name(info->name)
+            });
+        }
+    }
+    return devices;
+}
+
+bool AudioEngine::set_input_device(int device_index) {
+    const PaDeviceInfo* info = Pa_GetDeviceInfo(device_index);
+    if (!info || info->maxInputChannels < 1) {
+        last_error_ = "Invalid input device.";
+        return false;
+    }
+
+    if (!devices_share_host_api(device_index, output_device_)) {
+        const PaDeviceInfo* out_info = Pa_GetDeviceInfo(output_device_);
+        const PaHostApiInfo* in_api = Pa_GetHostApiInfo(info->hostApi);
+        const PaHostApiInfo* out_api = out_info ? Pa_GetHostApiInfo(out_info->hostApi) : nullptr;
+        std::cerr << "[Amplitron] Warning: Input (" << (in_api ? in_api->name : "?")
+                  << ") and output (" << (out_api ? out_api->name : "?")
+                  << ") are on different host APIs. Stream may fail." << std::endl;
+    }
+
+    int prev_device = input_device_;
+    bool was_running = running_;
+    if (was_running) stop();
+    input_device_ = device_index;
+    if (was_running) {
+        if (!start()) {
+            last_error_ = "Failed to start with new input device. Reverting.";
+            std::cerr << "[Amplitron] " << last_error_ << std::endl;
+            input_device_ = prev_device;
+            start();
+            return false;
+        }
+        last_error_.clear();
+    }
+    return true;
+}
+
+bool AudioEngine::set_output_device(int device_index) {
+    const PaDeviceInfo* info = Pa_GetDeviceInfo(device_index);
+    if (!info || info->maxOutputChannels < 1) {
+        last_error_ = "Invalid output device.";
+        return false;
+    }
+
+    if (!devices_share_host_api(input_device_, device_index)) {
+        const PaDeviceInfo* in_info = Pa_GetDeviceInfo(input_device_);
+        const PaHostApiInfo* in_api = in_info ? Pa_GetHostApiInfo(in_info->hostApi) : nullptr;
+        const PaHostApiInfo* out_api = Pa_GetHostApiInfo(info->hostApi);
+        std::cerr << "[Amplitron] Warning: Input (" << (in_api ? in_api->name : "?")
+                  << ") and output (" << (out_api ? out_api->name : "?")
+                  << ") are on different host APIs. Stream may fail." << std::endl;
+    }
+
+    int prev_device = output_device_;
+    bool was_running = running_;
+    if (was_running) stop();
+    output_device_ = device_index;
+    if (was_running) {
+        if (!start()) {
+            last_error_ = "Failed to start with new output device. Reverting.";
+            std::cerr << "[Amplitron] " << last_error_ << std::endl;
+            output_device_ = prev_device;
+            start();
+            return false;
+        }
+        last_error_.clear();
+    }
+    return true;
+}
+
+} // namespace Amplitron
