+++ src/audio/audio_backend_portaudio_internal.h
@@ -1,0 +1,15 @@
+#pragma once
+
+#include <portaudio.h>
+
+namespace Amplitron {
+
+/**
+ * @brief Opaque state for PortAudio backend.
+ * Defined here so all portaudio split files can access it.
+ */
+struct AudioBackendState {
+    PaStream* stream = nullptr;
+};
+
+} // namespace Amplitron
