+++ src/audio/recorder_impl.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <string>
+
+namespace Amplitron {
+
+// Forward-declared helper function, defined in recorder.cpp
+void mkdirs(const std::string& path);
+
+} // namespace Amplitron
