+++ src/gui/gui_manager.cpp
@@ -226,459 +226,5 @@ void GuiManager::shutdown() {
     SDL_Quit();
 }
 
-bool GuiManager::run_frame() {
-    SDL_Event event;
-    while (SDL_PollEvent(&event)) {
-        ImGui_ImplSDL2_ProcessEvent(&event);
-        if (event.type == SDL_QUIT) return false;
-        if (event.type == SDL_WINDOWEVENT &&
-            event.window.event == SDL_WINDOWEVENT_CLOSE &&
-            event.window.windowID == SDL_GetWindowID(window_))
-            return false;
-    }
-
-    // Poll MIDI events and apply CC mappings
-    midi_manager_.poll(engine_);
-
-    ImGui_ImplOpenGL3_NewFrame();
-    ImGui_ImplSDL2_NewFrame();
-    ImGui::NewFrame();
-
-    // Keyboard shortcuts for undo/redo and snapshot save
-    {
-        ImGuiIO& io = ImGui::GetIO();
-        if (!io.WantTextInput) {
-            bool mod = io.KeySuper || io.KeyCtrl;
-            if (mod && !io.KeyShift && ImGui::IsKeyPressed(ImGuiKey_Z)) {
-                if (command_history_.undo() && pedal_board_) {
-                    pedal_board_->rebuild_widgets();
-                }
-            }
-            if (mod && io.KeyShift && ImGui::IsKeyPressed(ImGuiKey_Z)) {
-                if (command_history_.redo() && pedal_board_) {
-                    pedal_board_->rebuild_widgets();
-                }
-            }
-            if (mod && !io.KeyShift && ImGui::IsKeyPressed(ImGuiKey_Y)) {
-                if (command_history_.redo() && pedal_board_) {
-                    pedal_board_->rebuild_widgets();
-                }
-            }
-            // Ctrl/Cmd+1–4: recall snapshot slot A–D
-            static const ImGuiKey digit_keys[4] = {
-                ImGuiKey_1, ImGuiKey_2, ImGuiKey_3, ImGuiKey_4
-            };
-            for (int i = 0; i < 4; ++i) {
-                if (mod && !io.KeyShift && ImGui::IsKeyPressed(digit_keys[i])) {
-                    if (gui_snapshots_.manager().has_slot(i)) {
-                        gui_snapshots_.recall_slot(i);
-                        if (pedal_board_) pedal_board_->rebuild_widgets();
-                    }
-                }
-            }
-        }
-    }
-
-    // Main menu bar
-    render_menu_bar();
-
-    // Full-window layout
-    SDL_GetWindowSize(window_, &window_width_, &window_height_);
-
-    ImGui::SetNextWindowPos(ImVec2(0, 20));
-    ImGui::SetNextWindowSize(ImVec2(static_cast<float>(window_width_),
-                                     static_cast<float>(window_height_) - 20));
-    ImGui::Begin("##MainArea", nullptr,
-        ImGuiWindowFlags_NoTitleBar | ImGuiWindowFlags_NoResize |
-        ImGuiWindowFlags_NoMove | ImGuiWindowFlags_NoCollapse |
-        ImGuiWindowFlags_NoBringToFrontOnFocus);
-
-    render_master_controls();
-
-    ImGui::Separator();
-
-    // Recording controls (above pedal board)
-    gui_recording_.render_controls();
-
-    ImGui::Separator();
-
-    // In-session snapshots (A/B/C/D slot row)
-    gui_snapshots_.render();
-
-    ImGui::Separator();
-
-    float analyzer_reserved_h = gui_analyzer_.analyzer_reserved_height();
-    ImGui::BeginChild("PedalBoardRegion", ImVec2(0, -analyzer_reserved_h), false);
-    if (pedal_board_) {
-        pedal_board_->render();
-    }
-    ImGui::EndChild();
-
-    ImGui::Separator();
-    gui_analyzer_.render();
-
-    ImGui::End();
-
-    // Popups / floating windows
-    if (show_settings_) {
-        gui_settings_.render(show_settings_);
-    }
-    if (show_save_preset_) {
-        gui_presets_.render_save_popup(show_save_preset_);
-    }
-    if (show_load_preset_) {
-        gui_presets_.render_load_popup(show_load_preset_);
-    }
-    if (gui_recording_.show_save()) {
-        gui_recording_.render_save_dialog(gui_recording_.show_save());
-    }
-    if (show_tuner_) {
-        gui_tuner_.render(show_tuner_);
-    }
-    if (show_midi_) {
-        gui_midi_.render(show_midi_);
-    }
-
-    // Rendering
-    ImGui::Render();
-    int display_w, display_h;
-    SDL_GL_GetDrawableSize(window_, &display_w, &display_h);
-    glViewport(0, 0, display_w, display_h);
-    glClearColor(0.078f, 0.071f, 0.063f, 1.0f);  // #141210 BG_DARKEST
-    glClear(GL_COLOR_BUFFER_BIT);
-    ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
-    SDL_GL_SwapWindow(window_);
-
-    return true;
-}
-
-void GuiManager::render_menu_bar() {
-    if (ImGui::BeginMainMenuBar()) {
-        if (ImGui::BeginMenu("File")) {
-            if (ImGui::MenuItem("New Preset...")) {
-                gui_presets_.begin_new_preset();
-                show_save_preset_ = true;
-            }
-            if (ImGui::MenuItem("Save Preset...", "Ctrl+S")) {
-                gui_presets_.begin_save_preset();
-                show_save_preset_ = true;
-            }
-            if (ImGui::MenuItem("Load Preset...", "Ctrl+O")) {
-                show_load_preset_ = true;
-                gui_presets_.ensure_factory_presets();
-                gui_presets_.refresh_presets(true);
-            }
-            bool has_selected_preset = gui_presets_.selected_preset_index() >= 0 &&
-                                       gui_presets_.selected_preset_index() < gui_presets_.preset_count();
-            if (ImGui::MenuItem("Delete Selected Preset", nullptr, false, has_selected_preset)) {
-                gui_presets_.delete_preset_by_index(gui_presets_.selected_preset_index());
-            }
-            ImGui::Separator();
-#ifndef AMPLITRON_NO_DESKTOP_SHELL
-            if (ImGui::MenuItem("Change Presets Directory...")) {
-                std::string chosen = show_folder_dialog("Select Presets Directory");
-                if (!chosen.empty()) {
-                    PresetManager::set_presets_dir(chosen);
-                    PresetManager::save_config();
-                    gui_presets_.refresh_presets(false);
-                }
-            }
-            if (ImGui::MenuItem("Reset to Default Presets Directory")) {
-                PresetManager::set_presets_dir("");
-                PresetManager::save_config();
-                gui_presets_.refresh_presets(false);
-            }
-#endif
-            ImGui::Separator();
-            if (ImGui::MenuItem("Settings")) show_settings_ = true;
-            ImGui::Separator();
-            if (ImGui::MenuItem("Quit", "Alt+F4")) {
-                SDL_Event quit_event;
-                quit_event.type = SDL_QUIT;
-                SDL_PushEvent(&quit_event);
-            }
-            ImGui::EndMenu();
-        }
-        if (ImGui::BeginMenu("Edit")) {
-            bool can_undo = command_history_.can_undo();
-            bool can_redo = command_history_.can_redo();
-
-            const char* undo_label = command_history_.undo_description();
-            char undo_buf[64] = "Undo";
-            if (undo_label) snprintf(undo_buf, sizeof(undo_buf), "Undo %s", undo_label);
-
-            const char* redo_label = command_history_.redo_description();
-            char redo_buf[64] = "Redo";
-            if (redo_label) snprintf(redo_buf, sizeof(redo_buf), "Redo %s", redo_label);
-
-            if (ImGui::MenuItem(undo_buf, "Ctrl+Z", false, can_undo)) {
-                if (command_history_.undo() && pedal_board_) {
-                    pedal_board_->rebuild_widgets();
-                }
-            }
-            if (ImGui::MenuItem(redo_buf, "Ctrl+Shift+Z", false, can_redo)) {
-                if (command_history_.redo() && pedal_board_) {
-                    pedal_board_->rebuild_widgets();
-                }
-            }
-            ImGui::EndMenu();
-        }
-        if (ImGui::BeginMenu("Audio")) {
-            if (engine_.is_running()) {
-                if (ImGui::MenuItem("Stop Audio")) engine_.stop();
-            } else {
-                if (ImGui::MenuItem("Start Audio")) {
-                    engine_.restart();
-                }
-            }
-            ImGui::Separator();
-            if (ImGui::MenuItem("Restart Audio")) {
-                engine_.restart();
-            }
-            ImGui::EndMenu();
-        }
-        if (ImGui::BeginMenu("Utilities")) {
-            if (ImGui::MenuItem("Open Tuner", nullptr, show_tuner_)) {
-                gui_tuner_.toggle(show_tuner_);
-            }
-            if (ImGui::MenuItem("MIDI Settings", nullptr, show_midi_)) {
-                show_midi_ = !show_midi_;
-            }
-            ImGui::EndMenu();
-        }
-
-        // Status bar (right side)
-        float bar_w = ImGui::GetWindowWidth();
-
-        // MIDI status indicator
-        ImGui::SameLine(bar_w - 450);
-        if (midi_manager_.is_port_open()) {
-            ImGui::TextColored(Theme::Live(), "MIDI");
-        } else {
-            ImGui::TextColored(ImVec4(0.4f, 0.4f, 0.4f, 1.0f), "MIDI");
-        }
-
-        // Recording indicator
-        ImGui::SameLine(bar_w - 400);
-        if (engine_.recorder().is_recording()) {
-            float t = static_cast<float>(ImGui::GetTime());
-            ImGui::TextColored(Theme::RecBlink(t), "REC");
-            ImGui::SameLine();
-            ImGui::Text("%.1fs", engine_.recorder().get_duration());
-        }
-
-        // Audio status
-        ImGui::SameLine(bar_w - 200);
-        if (engine_.is_running()) {
-            ImGui::TextColored(Theme::Live(), "LIVE");
-        } else {
-            ImGui::TextColored(Theme::Stopped(), "STOPPED");
-        }
-        ImGui::SameLine();
-        ImGui::Text("%dHz", engine_.get_sample_rate());
-
-        bool show_update = false;
-        std::string update_version;
-        std::string update_url;
-        {
-            std::lock_guard<std::mutex> lock(update_mutex_);
-            if (has_new_release_) {
-                show_update = true;
-                update_version = new_release_version_;
-                update_url = new_release_url_;
-            }
-        }
-
-        if (show_update) {
-            ImGui::SameLine(bar_w - 600);
-            ImGui::TextColored(Theme::GoldHot(), "New Release Available: %s", update_version.c_str());
-            if (ImGui::IsItemHovered()) {
-                ImGui::SetTooltip("Click to open release in browser");
-                ImGui::SetMouseCursor(ImGuiMouseCursor_Hand);
-            }
-            if (ImGui::IsItemClicked()) {
-#if defined(_WIN32)
-                std::string cmd = "start " + update_url;
-                std::system(cmd.c_str());
-#elif defined(__APPLE__) && !TARGET_OS_IOS
-                std::string cmd = "open " + update_url;
-                std::system(cmd.c_str());
-#elif defined(__linux__)
-                std::string cmd = "xdg-open " + update_url;
-                std::system(cmd.c_str());
-#endif
-            }
-        }
-
-        ImGui::EndMainMenuBar();
-    }
-
-    // Error banner when audio is stopped
-    if (!engine_.is_running()) {
-        ImGui::PushStyleColor(ImGuiCol_ChildBg, ImVec4(0.35f, 0.08f, 0.08f, 0.95f));
-        ImGui::BeginChild("AudioErrorBanner", ImVec2(0, 36), true);
-        ImGui::TextColored(Theme::Stopped(), "Audio stream is STOPPED.");
-        ImGui::SameLine();
-        if (ImGui::SmallButton("Restart Audio")) {
-            engine_.restart();
-        }
-        ImGui::SameLine();
-        if (ImGui::SmallButton("Settings")) {
-            show_settings_ = true;
-        }
-        std::string err = engine_.get_last_error();
-        if (!err.empty()) {
-            ImGui::SameLine();
-            ImGui::TextColored(Theme::GoldHot(), "  %s", err.c_str());
-        }
-        ImGui::EndChild();
-        ImGui::PopStyleColor();
-    }
-}
-
-void GuiManager::render_master_controls() {
-    // Smooth metering
-    float input_lvl = engine_.get_input_level();
-    float output_lvl = engine_.get_output_level();
-    smoothed_input_level_ += (input_lvl - smoothed_input_level_) * 0.3f;
-    smoothed_output_level_ += (output_lvl - smoothed_output_level_) * 0.3f;
-
-    ImGui::BeginChild("MasterControls", ImVec2(0, 80), true);
-
-    ImGui::Columns(4, "master_cols", false);
-
-    // Input gain
-    ImGui::Text("INPUT");
-    float input_gain = engine_.get_input_gain();
-    if (ImGui::SliderFloat("##InputGain", &input_gain, 0.0f, 5.0f, "%.2f")) {
-        engine_.set_input_gain(input_gain);
-    }
-
-    ImGui::NextColumn();
-
-    // Input meter
-    ImGui::Text("IN LEVEL");
-    ImVec2 meter_pos = ImGui::GetCursorScreenPos();
-    float meter_w = ImGui::GetColumnWidth() - 20;
-    ImDrawList* dl = ImGui::GetWindowDrawList();
-    dl->AddRectFilled(meter_pos, ImVec2(meter_pos.x + meter_w, meter_pos.y + 20),
-                       Theme::METER_BG, Theme::ROUNDING_SM);
-    float fill = std::min(smoothed_input_level_, 1.0f) * meter_w;
-    ImU32 meter_color = (smoothed_input_level_ > 0.9f) ? Theme::METER_RED :
-                        (smoothed_input_level_ > 0.6f) ? Theme::METER_YELLOW :
-                                                          Theme::METER_GREEN;
-    dl->AddRectFilled(meter_pos, ImVec2(meter_pos.x + fill, meter_pos.y + 20),
-                       meter_color, Theme::ROUNDING_SM);
-    ImGui::Dummy(ImVec2(meter_w, 20));
-
-    ImGui::NextColumn();
-
-    // Output meter
-    ImGui::Text("OUT LEVEL");
-    meter_pos = ImGui::GetCursorScreenPos();
-    meter_w = ImGui::GetColumnWidth() - 20;
-    dl->AddRectFilled(meter_pos, ImVec2(meter_pos.x + meter_w, meter_pos.y + 20),
-                       Theme::METER_BG, Theme::ROUNDING_SM);
-    fill = std::min(smoothed_output_level_, 1.0f) * meter_w;
-    meter_color = (smoothed_output_level_ > 0.9f) ? Theme::METER_RED :
-                  (smoothed_output_level_ > 0.6f) ? Theme::METER_YELLOW :
-                                                     Theme::METER_GREEN;
-    dl->AddRectFilled(meter_pos, ImVec2(meter_pos.x + fill, meter_pos.y + 20),
-                       meter_color, Theme::ROUNDING_SM);
-    ImGui::Dummy(ImVec2(meter_w, 20));
-
-    ImGui::NextColumn();
-
-    // Output gain
-    ImGui::Text("OUTPUT");
-    float output_gain = engine_.get_output_gain();
-    if (ImGui::SliderFloat("##OutputGain", &output_gain, 0.0f, 2.0f, "%.2f")) {
-        engine_.set_output_gain(output_gain);
-    }
-
-    ImGui::Columns(1);
-    ImGui::EndChild();
-}
-
-void GuiManager::check_for_updates() {
-#ifndef AMPLITRON_NO_DESKTOP_SHELL
-    FILE* pipe = nullptr;
-#ifdef _WIN32
-    pipe = _popen("curl -s https://api.github.com/repos/sudip-mondal-2002/Amplitron/releases", "r");
-#else
-    pipe = popen("curl -s https://api.github.com/repos/sudip-mondal-2002/Amplitron/releases", "r");
-#endif
-
-    if (!pipe) return;
-
-    std::string result = "";
-    char buffer[256];
-    while (!feof(pipe)) {
-        if (fgets(buffer, 256, pipe) != nullptr)
-            result += buffer;
-    }
-#ifdef _WIN32
-    _pclose(pipe);
-#else
-    pclose(pipe);
-#endif
-
-    std::string search_str = "\"tag_name\": \"";
-    size_t pos = result.find(search_str);
-    if (pos != std::string::npos) {
-        pos += search_str.length();
-        size_t end_pos = result.find("\"", pos);
-        if (end_pos != std::string::npos) {
-            std::string latest_version = result.substr(pos, end_pos - pos);
-
-            std::string html_url = "";
-            std::string url_search_str = "\"html_url\": \"";
-            size_t url_pos = result.find(url_search_str);
-            if (url_pos != std::string::npos) {
-                url_pos += url_search_str.length();
-                size_t url_end_pos = result.find("\"", url_pos);
-                if (url_end_pos != std::string::npos) {
-                    html_url = result.substr(url_pos, url_end_pos - url_pos);
-                }
-            }
-
-            auto parse_version = [](const std::string& v) -> std::vector<int> {
-                std::vector<int> parts;
-                std::string s = v;
-                if (!s.empty() && s[0] == 'v') s = s.substr(1);
-                size_t pos = 0;
-                while (pos < s.size()) {
-                    size_t dot = s.find('.', pos);
-                    if (dot == std::string::npos) dot = s.size();
-                    try { parts.push_back(std::stoi(s.substr(pos, dot - pos))); }
-                    catch (...) { parts.push_back(0); }
-                    pos = dot + 1;
-                }
-                return parts;
-            };
-
-            std::string current_version = "v" AMPLITRON_VERSION;
-            if (!latest_version.empty()) {
-                auto latest_parts = parse_version(latest_version);
-                auto current_parts = parse_version(current_version);
-                bool is_newer = false;
-                size_t max_len = std::max(latest_parts.size(), current_parts.size());
-                for (size_t i = 0; i < max_len; ++i) {
-                    int lv = (i < latest_parts.size()) ? latest_parts[i] : 0;
-                    int cv = (i < current_parts.size()) ? current_parts[i] : 0;
-                    if (lv > cv) { is_newer = true; break; }
-                    if (lv < cv) { break; }
-                }
-                if (is_newer) {
-                    std::lock_guard<std::mutex> lock(update_mutex_);
-                    new_release_version_ = latest_version;
-                    new_release_url_ = html_url;
-                    has_new_release_ = true;
-                }
-            }
-        }
-    }
-#endif
-}
 
 } // namespace Amplitron
