+++ src/gui/pedal_board.cpp
@@ -2,28 +2,11 @@
 #include "gui/pedal_widget.h"
 #include "gui/theme.h"
 #include "gui/command.h"
-
-#include "audio/effects/noise_gate.h"
-#include "audio/effects/compressor.h"
-#include "audio/effects/overdrive.h"
-#include "audio/effects/distortion.h"
-#include "audio/effects/equalizer.h"
-#include "audio/effects/chorus.h"
-#include "audio/effects/phaser.h"
-#include "audio/effects/flanger.h"
-#include "audio/effects/delay.h"
-#include "audio/effects/reverb.h"
-#include "audio/effects/cabinet_sim.h"
-#include "audio/effects/ir_cabinet.h"
 #include "audio/effects/amp_simulator.h"
-#include "audio/effects/wah.h"
-#include "audio/effects/octaver.h"
-#include "audio/effects/pitch_shifter.h"
-#include <cstring>
-#include <cstdio>
-#include <set>
 
+#include <cstring>
 #include <imgui.h>
+#include <set>
 
 namespace Amplitron {
 
@@ -147,232 +130,5 @@ void PedalBoard::add_effect_and_show(std::shared_ptr<Effect> effect) {
     rebuild_widgets();
 }
 
-/** @brief Render the "+ Add Pedal" button and category popup with effect menu items.
- *  Amps and Tuner are handled separately (amp selector dropdown, tuner modal). */
-void PedalBoard::render_add_pedal_menu() {
-    if (ImGui::Button("+ Add Pedal")) {
-        ImGui::OpenPopup("AddPedalPopup");
-    }
-
-    if (ImGui::BeginPopup("AddPedalPopup")) {
-        ImGui::TextColored(Theme::Gold(), "DRIVE");
-        if (ImGui::MenuItem("Overdrive")) {
-            add_effect_and_show(std::make_shared<Overdrive>());
-        }
-        if (ImGui::MenuItem("Distortion")) {
-            add_effect_and_show(std::make_shared<Distortion>());
-        }
-
-        ImGui::Separator();
-        ImGui::TextColored(Theme::Live(), "DYNAMICS");
-        if (ImGui::MenuItem("Noise Gate")) {
-            add_effect_and_show(std::make_shared<NoiseGate>());
-        }
-        if (ImGui::MenuItem("Compressor")) {
-            add_effect_and_show(std::make_shared<Compressor>());
-        }
-
-        ImGui::Separator();
-        ImGui::TextColored(ImVec4(0.35f, 0.60f, 0.95f, 1.0f), "MODULATION");
-        if (ImGui::MenuItem("Chorus")) {
-            add_effect_and_show(std::make_shared<Chorus>());
-        }
-        if (ImGui::MenuItem("Phaser")) {
-            add_effect_and_show(std::make_shared<Phaser>());
-        }
-        if (ImGui::MenuItem("Flanger")) {
-            add_effect_and_show(std::make_shared<Flanger>());
-        }
-
-        ImGui::Separator();
-        ImGui::TextColored(ImVec4(0.65f, 0.35f, 0.95f, 1.0f), "TIME");
-        if (ImGui::MenuItem("Delay")) {
-            add_effect_and_show(std::make_shared<Delay>());
-        }
-        if (ImGui::MenuItem("Reverb")) {
-            add_effect_and_show(std::make_shared<Reverb>());
-        }
-
-        ImGui::Separator();
-        ImGui::TextColored(ImVec4(0.30f, 0.75f, 0.60f, 1.0f), "FILTER");
-        if (ImGui::MenuItem("Wah")) {
-            add_effect_and_show(std::make_shared<WahPedal>());
-        }
-
-        ImGui::Separator();
-        ImGui::TextColored(ImVec4(0.85f, 0.40f, 0.55f, 1.0f), "PITCH");
-        if (ImGui::MenuItem("Octaver")) {
-            add_effect_and_show(std::make_shared<Octaver>());
-        }
-        if (ImGui::MenuItem("Pitch Shifter")) {
-            add_effect_and_show(std::make_shared<PitchShifter>());
-        }
-
-        ImGui::Separator();
-        ImGui::TextColored(Theme::GoldDim(), "TONE");
-        if (ImGui::MenuItem("Equalizer")) {
-            add_effect_and_show(std::make_shared<Equalizer>());
-        }
-        if (ImGui::MenuItem("Cabinet Sim")) {
-            add_effect_and_show(std::make_shared<CabinetSim>());
-        }
-        if (ImGui::MenuItem("IR Cabinet")) {
-            add_effect_and_show(std::make_shared<IRCabinet>());
-        }
-
-        ImGui::EndPopup();
-    }
-}
-
-/** @brief Render the amp model selector dropdown (max 1 amp, always present). */
-void PedalBoard::render_amp_selector() {
-    const auto& models = get_amp_models();
-    int amp_idx = find_amp_index();
-
-    // Auto-add default amp if none exists
-    if (amp_idx < 0) {
-        auto amp = std::make_shared<AmpSimulator>();
-        amp->params()[0].value = 0.0f; // first model
-        engine_.add_effect(amp);
-        rebuild_widgets();
-        amp_idx = find_amp_index();
-    }
-
-    // Current model label
-    const char* current_label = "Amp";
-    int current_model = 0;
-    if (amp_idx >= 0) {
-        auto& amp_fx = engine_.effects()[amp_idx];
-        current_model = static_cast<int>(amp_fx->params()[0].value);
-        if (current_model >= 0 && current_model < static_cast<int>(models.size())) {
-            current_label = models[current_model].name;
-        }
-    }
-
-    ImGui::PushStyleColor(ImGuiCol_Button, ImVec4(0.35f, 0.18f, 0.08f, 1.0f));
-    ImGui::PushStyleColor(ImGuiCol_ButtonHovered, ImVec4(0.50f, 0.25f, 0.10f, 1.0f));
-    char amp_label[64];
-    std::snprintf(amp_label, sizeof(amp_label), "Amp: %s", current_label);
-    if (ImGui::Button(amp_label)) {
-        ImGui::OpenPopup("AmpSelectorPopup");
-    }
-    ImGui::PopStyleColor(2);
-
-    if (ImGui::BeginPopup("AmpSelectorPopup")) {
-        ImGui::TextColored(ImVec4(0.95f, 0.55f, 0.20f, 1.0f), "AMP MODEL");
-        for (int m = 0; m < static_cast<int>(models.size()); ++m) {
-            bool is_selected = (current_model == m);
-            if (ImGui::MenuItem(models[m].name, models[m].inspiration, is_selected)) {
-                if (amp_idx >= 0) {
-                    engine_.effects()[amp_idx]->params()[0].value = static_cast<float>(m);
-                }
-            }
-        }
-        ImGui::EndPopup();
-    }
-}
-
-/** @brief Draw the signal flow line, render each pedal widget, and handle drag-and-drop reordering.
- *  Uses visibility set to determine which pedals to show. */
-void PedalBoard::render_signal_chain() {
-    // Build list of visible widget indices from visibility set
-    std::vector<int> visible;
-    for (int idx : visible_indices_) {
-        if (idx >= 0 && idx < static_cast<int>(widgets_.size())) {
-            visible.push_back(idx);
-        }
-    }
-
-    if (visible.empty()) {
-        ImGui::SetCursorPos(ImVec2(
-            ImGui::GetWindowWidth() / 2 - 150,
-            ImGui::GetWindowHeight() / 2 - 30
-        ));
-        ImGui::TextColored(Theme::TextDim(),
-            "No pedals in chain.\nClick '+ Add Pedal' to get started.");
-        return;
-    }
-
-    ImDrawList* dl = ImGui::GetWindowDrawList();
-    ImVec2 origin = ImGui::GetCursorScreenPos();
-
-    // Draw signal flow line
-    float line_y = origin.y + 160;
-    float total_width = visible.size() * 195.0f + 40;
-    dl->AddLine(
-        ImVec2(origin.x, line_y),
-        ImVec2(origin.x + total_width, line_y),
-        Theme::CHAIN_LINE, 3.0f
-    );
-
-    // Input jack
-    dl->AddCircleFilled(ImVec2(origin.x + 5, line_y), 6, Theme::CHAIN_JACK);
-    dl->AddCircle(ImVec2(origin.x + 5, line_y), 6, Theme::BORDER_DARK, 0, 1.5f);
-
-    // Render each visible pedal
-    float pedal_x = origin.x + 20;
-    int remove_idx = -1;
-
-    for (int vi = 0; vi < static_cast<int>(visible.size()); ++vi) {
-        int i = visible[vi];
-        ImVec2 pedal_min = ImVec2(pedal_x, origin.y + 5);
-
-        // Drag-and-drop reordering — render the full-pedal hit area FIRST as a
-        // background layer, then allow the widget's knobs/switches to overlap it.
-        ImGui::SetCursorScreenPos(pedal_min);
-        char dnd_id[32];
-        snprintf(dnd_id, sizeof(dnd_id), "##dnd_%d", i);
-        ImGui::SetNextItemAllowOverlap();
-        ImGui::InvisibleButton(dnd_id, ImVec2(Theme::PEDAL_WIDTH, Theme::PEDAL_HEIGHT));
-
-        bool is_amp = std::strcmp(widgets_[i]->get_effect()->name(), "Amp Sim") == 0;
-        if (!is_amp) {
-            if (ImGui::BeginDragDropSource(ImGuiDragDropFlags_SourceAllowNullID)) {
-                ImGui::SetDragDropPayload("PEDAL_REORDER", &i, sizeof(int));
-                ImGui::Text("Move %s", widgets_[i]->get_effect()->name());
-                ImGui::EndDragDropSource();
-            }
-            if (ImGui::BeginDragDropTarget()) {
-                if (const ImGuiPayload* payload = ImGui::AcceptDragDropPayload("PEDAL_REORDER")) {
-                    int source_idx = *static_cast<const int*>(payload->Data);
-                    if (source_idx != i) {
-                        history_.execute(std::make_unique<ReorderEffectCommand>(engine_, source_idx, i));
-                        rebuild_widgets();
-                    }
-                }
-                ImGui::EndDragDropTarget();
-            }
-        }
-
-        // Render the pedal widget on top — its interactive elements take priority
-        // over the DND background button thanks to SetNextItemAllowOverlap above.
-        ImGui::SetCursorScreenPos(pedal_min);
-        if (widgets_[i]->render()) {
-            remove_idx = i;
-        }
-
-        // Connection dot between pedals
-        if (vi < static_cast<int>(visible.size()) - 1) {
-            float dot_x = pedal_x + 190;
-            dl->AddCircleFilled(ImVec2(dot_x, line_y), 4, Theme::CHAIN_DOT);
-        }
-
-        pedal_x += 195;
-    }
-
-    // Output jack
-    dl->AddCircleFilled(ImVec2(pedal_x, line_y), 6, Theme::CHAIN_JACK);
-    dl->AddCircle(ImVec2(pedal_x, line_y), 6, Theme::BORDER_DARK, 0, 1.5f);
-
-    // Handle removal
-    if (remove_idx >= 0) {
-        visible_indices_.erase(remove_idx);
-        history_.execute(std::make_unique<RemoveEffectCommand>(engine_, remove_idx));
-        rebuild_widgets();
-    }
-
-    // Reserve space for horizontal scrolling
-    ImGui::Dummy(ImVec2(total_width + 20, 340));
-}
 
 } // namespace Amplitron
